requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
requests==2.31.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-dotenv==1.0.0
cohere==4.37
